backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
prisma==0.11.0
python-dotenv==1.0.0
Pillow==10.1.0
bcrypt==4.1.2
